backend/requirements.txt
asgiref==3.11.1
chromadb
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
google-genai
openai
pymysql
python-dotenv
rank-bm25
sqlparse==0.5.5
tzdata==2026.1
